Adding empty class structure for devices & factory method
This covers all of the device types that I think will be commonly used. There are many subcategories of open/close and motion sensor but I am making an assumption that they all have the same basic functionality and packet structure. If that proves to be wrong, we can add specific subcat classes for these devices later.